script fix

+++ install-host.sh
@@ -25,9 +25,10 @@ apt-get -y install software-properties-common byobu zsh git htop mc
 # LXC
 add-apt-repository ppa:ubuntu-lxc/daily
 apt-get install lxc
-ln -s /var/lib/lxc /lxc
+if [[ ! -e /lxc ]]; then ln -s /var/lib/lxc /lxc; fi
 
 # dotfiles & others
+mkdir -p $DIR
 git clone https://github.com/niieani/lxc-shared.git ${DIR}
 (cd /shared; git submodule init && git submodule update && git submodule status)
